+++ ts/features/payments/common/api/client.ts
@@ -1,7 +1,8 @@
 import { createClient } from "../../../../../definitions/pagopa/walletv3/client";
+import { createClient as createECommerceClient } from "../../../../../definitions/pagopa/ecommerce/client";
 import { defaultRetryingFetch } from "../../../../utils/fetch";
 
-const createWalletClient = (baseUrl: string, bearerAuth: string) =>
+export const createWalletClient = (baseUrl: string, bearerAuth: string) =>
   createClient<"bearerAuth">({
     baseUrl,
     basePath: "/payment-wallet/v1",
@@ -16,6 +17,20 @@ const createWalletClient = (baseUrl: string, bearerAuth: string) =>
     }
   });
 
-export type WalletClient = ReturnType<typeof createWalletClient>;
+export const createPaymentClient = (baseUrl: string, token: string) =>
+  createECommerceClient<"walletToken">({
+    baseUrl,
+    basePath: "/ecommerce/io/v1",
+    fetchApi: defaultRetryingFetch(),
+    withDefaults: op => params => {
+      const paramsWithDefaults = {
+        ...params,
+        walletToken: token
+      } as Parameters<typeof op>[0];
 
-export { createWalletClient };
+      return op(paramsWithDefaults);
+    }
+  });
+
+export type PaymentClient = ReturnType<typeof createPaymentClient>;
+export type WalletClient = ReturnType<typeof createWalletClient>;
