+++ ts/features/payments/common/saga/index.ts
@@ -2,9 +2,8 @@ import { SagaIterator } from "redux-saga";
 import { fork, select } from "typed-redux-saga/macro";
 import { isPagoPATestEnabledSelector } from "../../../../store/reducers/persistedPreferences";
 import { watchPaymentsOnboardingSaga } from "../../onboarding/saga";
-import { createPaymentClient } from "../../payment/api/client";
 import { watchPaymentsPaymentSaga } from "../../payment/saga";
-import { createWalletClient } from "../api/client";
+import { createPaymentClient, createWalletClient } from "../api/client";
 import { walletApiBaseUrl, walletApiUatBaseUrl } from "../../../../config";
 import { watchPaymentsMethodDetailsSaga } from "../../details/saga";
 import { watchPaymentsTransactionSaga } from "../../transaction/saga";
