+++ ts/features/payments/payment/saga/networking/handleWalletPaymentNewSessionToken.ts
@@ -6,7 +6,7 @@ import { SagaCallReturnType } from "../../../../../types/utils";
 import { getGenericError, getNetworkError } from "../../../../../utils/errors";
 import { readablePrivacyReport } from "../../../../../utils/reporters";
 import { withRefreshApiCall } from "../../../../fastLogin/saga/utils";
-import { PaymentClient } from "../../api/client";
+import { PaymentClient } from "../../../common/api/client";
 import { paymentsGetNewSessionTokenAction } from "../../store/actions/networking";
 import { selectWalletPaymentSessionToken } from "../../store/selectors";
